+++ oneview/provider.go
@@ -78,6 +78,7 @@ func Provider() terraform.ResourceProvider {
 			"oneview_scope":                      dataSourceScope(),
 			"oneview_server_hardware":            dataSourceServerHardware(),
 			"oneview_server_hardware_type":       dataSourceServerHardwareType(),
+			"oneview_storage_attachment":         dataSourceStorageAttachment(),
 			"oneview_logical_enclosure":          dataSourceLogicalEnclosure(),
 			"oneview_enclosure_group":            dataSourceEnclosureGroup(),
 			"oneview_server_profile":             dataSourceServerProfile(),
@@ -85,6 +86,7 @@ func Provider() terraform.ResourceProvider {
 			"oneview_storage_system":             dataSourceStorageSystem(),
 			"oneview_uplink_set":                 dataSourceUplinkSet(),
 			"oneview_network_set":                dataSourceNetworkSet(),
+			"oneview_storage_volume_template":    dataSourceStorageVolumeTemplate(),
 			"oneview_volume":                     dataSourceVolume(),
 		},
 
@@ -107,6 +109,7 @@ func Provider() terraform.ResourceProvider {
 			"oneview_i3s_plan":                   resourceI3SPlan(),
 			"oneview_logical_enclosure":          resourceLogicalEnclosure(),
 			"oneview_storage_pool":               resourceStoragePool(),
+			"oneview_storage_volume_template":    resourceStorageVolumeTemplate(),
 			"oneview_volume":                     resourceVolume(),
 		},
 		ConfigureFunc: providerConfigure,